requirements.txt
# 核心深度學習框架
torch
torchvision
numpy

# 圖像處理
Pillow

# API 服務與後端
fastapi
uvicorn
python-multipart

# Triton Client (包含 HTTP 支援)
tritonclient[http]
